+++ connect/common/signals.py
@@ -35,6 +35,7 @@ logger = logging.getLogger("connect.common.signals")
 
 @receiver(post_save, sender=Project)
 def create_service_status(sender, instance, created, **kwargs):
+    update_fields = list(kwargs.get("update_fields")) if kwargs.get("update_fields") else None
     if created:
         for service in Service.objects.filter(default=True):
             instance.service_status.create(service=service)
@@ -62,22 +63,15 @@ def create_service_status(sender, instance, created, **kwargs):
                 )
                 instance.created_by.send_request_flow_user_info(data)
 
-        for permission in instance.project_authorizations.all():
-            update_user_permission_project(
-                flow_organization=str(instance.flow_organization),
-                project_uuid=str(instance.uuid),
-                user_email=permission.user.email,
-                permission=permission.role
-            )
-            # celery_app.send_task(
-            #     "update_user_permission_project",
-            #     args=[
-            #         instance.flow_organization,
-            #         instance.uuid,
-            #         permission.user.email,
-            #         permission.role,
-            #     ],
-            # )
+        if instance.flow_organization:
+            for permission in instance.project_authorizations.all():
+                update_user_permission_project(
+                    project_uuid=str(instance.uuid),
+                    flow_organization=str(instance.flow_organization),
+                    user_email=permission.user.email,
+                    permission=permission.role
+                )
+
         for authorization in instance.organization.authorizations.all():
             if authorization.can_contribute:
                 project_auth = instance.get_user_authorization(authorization.user)
@@ -90,6 +84,14 @@ def create_service_status(sender, instance, created, **kwargs):
                         project=project_auth.project,
                         created_by=project_auth.user
                     )
+    elif "flow_organization" in update_fields:
+        for permission in instance.project_authorizations.all():
+            update_user_permission_project(
+                project_uuid=str(instance.uuid),
+                flow_organization=str(instance.flow_organization),
+                user_email=permission.user.email,
+                permission=permission.role
+            )
 
 
 @receiver(post_save, sender=Service)
@@ -105,7 +107,7 @@ def update_organization(instance, **kwargs):
         celery_app.send_task(  # pragma: no cover
             name="update_suspend_project",
             args=[
-                str(project.flow_organization),
+                str(project.uuid),
                 instance.is_suspended,
             ],
         )
@@ -292,22 +294,13 @@ def project_authorization(sender, instance, created, **kwargs):
         if instance_user.level == OrganizationLevelRole.NOTHING.value:
             instance_user.role = OrganizationRole.VIEWER.value
             instance_user.save(update_fields=["role"])
-
-        update_user_permission_project(
-            flow_organization=str(instance.project.flow_organization),
-            project_uuid=str(instance.project.uuid),
-            user_email=instance.user.email,
-            permission=instance.role
-        )
-        # celery_app.send_task(
-        #     "update_user_permission_project",
-        #     args=[
-        #         instance.project.flow_organization,
-        #         instance.project.uuid,
-        #         instance.user.email,
-        #         instance.role,
-        #     ],
-        # )
+        if instance.project.flow_organization:
+            update_user_permission_project(
+                project_uuid=str(instance.project.uuid),
+                flow_organization=str(instance.project.flow_organization),
+                user_email=instance.user.email,
+                permission=instance.role
+            )
 
 
 @receiver(post_save, sender=RequestRocketPermission)