+++ CHANGELOG.md
@@ -1,6 +1,64 @@
+# 2.10.0
+
+## *Add*
+  - New project template
+  - Verify user after 2nd login
+
+# 2.9.0
+## *Fix*
+  - template email translations
+  - Delete organization newsletter after changing plan
+## *Add*
+  - Endpoint to get user info about organizations
+
+# 2.8.1
+## *Fix*
+  - fix api_v2_urls import
+
+# 2.8.0
+## *Change*
+  - Connect-Flows Communication; Passing `project.uuid` instead of `project.flow_organization`
+
+# 2.7.2
+## *Fix*
+  - Increase timeout for get_project_statistic request
+
+# 2.7.1
+## *Fix*
+  - Coverage Files
+
+# 2.7.0
+## *Add*
+  - Template_project APP
+  - Omie Template
+
+# 2.5.0
+## *Add*
+  - NewsletterOrganization api
+
+# 2.4.0
+
+## *Add*
+  - WeniAI external proxy server
+
+# 2.2.0
+## *Add*
+  - New send email method
+  - New template messages
+
+# 2.1.0
+
+## *Update*
+  - Readme.md keycloack information
+    - Enviroments variables needed for running keycloak
+
+# 2.0.0
+## *Add*
+  - Updating routers for v2 api urls
+
 # 1.2.1
 ## Change
-  - Change `SyncManagerTask filter` in `count_contacts`  to only get objects with `status=True` 
+  - Change `SyncManagerTask filter` in `count_contacts`  to only get objects with `status=True`
   - remove `task.wait()` from `retry_billing_tasks`
   - Replaces Message `get_or_create` method with `try/except`
 
@@ -33,6 +91,6 @@
 
 ## DELETE:
   - the send emails call needed with the design review and user experience
-  
+
 ## CHANGE:
   - An organization admin can't be removed of an organization project